+++ js/main.js
@@ -1,8 +1,11 @@
 // eslint-disable-next-line no-redeclare,no-unused-vars
 /* global Pristine:readonly */
+// eslint-disable-next-line no-redeclare,no-unused-vars
+/* global noUiSlider:readonly */
 import { drawsThumbnails } from './draws-thumbnails.js';
 import './upload-file-modal.js';
 import './open-full-size-photo.js';
 import './validation.js';
+import './filters-user-photo.js';
 
 drawsThumbnails();
